% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid-subset-selection-base.R
\name{runHSS}
\alias{runHSS}
\title{runHSS}
\usage{
runHSS(x, y, score.method, custom.score.method = NULL)
}
\arguments{
\item{x}{table with predictors of interest}

\item{y}{vector of class labels}

\item{score.method}{scoring metric to use to perform HSS}

\item{custom.score.method}{function for your custom scoring metric.}
}
\description{
This function runs hybrid subset selection.
}
\keyword{internal}
